Renamed component to Icon instead of Index.
