+++ app/react/UI/Icon/Icon.js
@@ -1,6 +1,4 @@
 import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-export const Icon = (props) => {
-  return <FontAwesomeIcon {...props} />
-}
+export default props => <FontAwesomeIcon {...props} />;
