+++ app/react/UI/index.js
@@ -1,1 +1,5 @@
-export { Icon } from './Icon';
+import Icon from './Icon/Icon';
+
+export default {
+  Icon,
+};
